Revert "Revert "RavenDB-18379 - Separate ServerWide pin to mentor node tests to another test with 'DoNotReuseServer'
